extract disappearing messages worker to a separate file
